refactored withStateHandlers HOC to plain wrapper component
