+++ package.json
@@ -28,7 +28,6 @@
     "invariant": "^2.2.4",
     "lodash": "^4.17.4",
     "prop-types": "^15.6.0",
-    "recompose": "^0.30.0",
     "requestidlecallback": "^0.3.0",
     "substyle": "^6.1.2"
   },
