--- src/higher-order/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default as autoPositionHandling } from './autoPositionHandling'
